server: export strm for testing
